+++ lib/features/post_list/presentation/providers/post_providers.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:sodong_app/features/post_list/data/datasources/post_cache_data_source.dart';
+import 'package:sodong_app/features/post/data/data_source/post_cache_data_source.dart';
 import 'package:sodong_app/features/post_list/data/datasources/post_remote_data_source.dart';
 import 'package:sodong_app/features/post_list/data/repositories/post_repository_impl.dart';
 import 'package:sodong_app/features/post_list/domain/repositories/post_repository.dart';
